Creates a favorite button file and implements the collectionView Datasourceon a different file
